MXS-4761 Configure 'samples' and 'percentile'
Also ensure that we do not end up dividing by 0, in case of
patholigical samples and/or percentile.

+++ server/modules/routing/diff/diffconfig.hh
@@ -59,7 +59,7 @@ public:
     int64_t                    max_execution_time_difference;
     int64_t                    max_request_lag;
     mxs::config::Enum<OnError> on_error;
-    double                     percentile;
+    int64_t                    percentile;
     std::chrono::milliseconds  period;
     mxs::config::Enum<Report>  report;
     bool                       reset_replication;
